[0x0C ~ 2-calloc.c] (fix: add the memory is set to zero)

+++ 0x0C-more_malloc_free/2-calloc.c
@@ -11,7 +11,7 @@ void *_calloc(unsigned int nmemb, unsigned int size)
 {
 	void *buffer;
 
-	int long i;
+	unsigned int i;
 
 	if (nmemb == 0 || size == 0)
 		return (NULL);
